back-end/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
instructor
